+++ CHANGELOG.md
@@ -47,6 +47,8 @@ This project adheres to [Semantic Versioning](https://semver.org/).
    * Improved robustness of parsing delimited files.
    * Updated `pysat.Instrument.pandas_format` so that changes to setting
      are immediately reflected by changes to Instrument object.
+   * Added a test function for bad class/function/method input to reduce
+     duplicate code and improve message test combliance
 * Deprecations
    * Removed `freq` as a standard kwarg for `pysat.Instruments.download`
    * Deprecated `_test_download_travis` as a standard attribute for